Модель MediaFile перенесена в приложение services (переименованна согласно PEP8 CamelCase) Создана модель Portfolio
Рефакторинг связанных импортов
