+++ api/serializers.py
@@ -2,8 +2,8 @@ from rest_framework import serializers
 
 from orders.models import Chat, Message, Order, Raiting
 from properties.models import Feedback_property, Property, Room
-from services.models import Service
-from users.models import Media_file, User
+from services.models import Service, MediaFile
+from users.models import User
 
 
 class UserSerializer(serializers.ModelSerializer):
@@ -74,7 +74,7 @@ class FBpropertySerializer(serializers.ModelSerializer):
 
 class MediafileSerializer(serializers.ModelSerializer):
     class Meta:
-        model = Media_file
+        model = MediaFile
         fields = (
             'link',
             'title',
