+++ api/views.py
@@ -4,8 +4,8 @@ from rest_framework import viewsets
 from api.paginators import LimitPageNumberPagination
 from orders.models import Chat, Message, Order, Raiting
 from properties.models import Feedback_property, Property, Room
-from services.models import Service
-from users.models import Media_file, User
+from services.models import Service, MediaFile
+from users.models import User
 
 from .serializers import (
     ChatSerializer,
@@ -26,7 +26,7 @@ class UserViewSet(DjoserUserViewSet):
 
 
 class MediafileViewSet(viewsets.ModelViewSet):
-    queryset = Media_file.objects.all()
+    queryset = MediaFile.objects.all()
     serializer_class = MediafileSerializer
